+++ src/views/index.ts
@@ -5,3 +5,5 @@
 export { SquadTreeProvider, SquadTreeItem } from './SquadTreeProvider';
 export { WorkDetailsWebview } from './WorkDetailsWebview';
 export { IssueDetailWebview } from './IssueDetailWebview';
+export { SquadStatusBar } from './SquadStatusBar';
+export { SquadDashboardWebview } from './SquadDashboardWebview';